+++ save/.json
@@ -1,1 +1,1 @@
-{"gamePanelData":{"monsterArr":[{"x":1382,"y":1830,"type":"Monster"},{"x":1382,"y":1830,"type":"Monster"},{"x":1382,"y":1830,"type":"Monster"},{"x":1382,"y":1830,"type":"Monster"}],"elementArr":[{"x":480,"y":528,"type":"Door"},{"x":1104,"y":336,"type":"PowerUp"},{"x":912,"y":1776,"type":"Fire"},{"x":912,"y":1728,"type":"Fire"},{"x":912,"y":1872,"type":"Fire"},{"x":912,"y":1920,"type":"Fire"}],"npcArr":[{"x":1182,"y":280,"type":"Villager"}]},"playerData":{"spriteCounter":11,"worldX":536,"lives":6,"worldY":2007,"spriteNum":1,"speed":5,"direction":1,"rubies":5}}
+{"gamePanelData":{"monsterArr":[{"x":1382,"y":1830,"type":"Monster"},{"x":678,"y":1142,"type":"Monster"},{"x":1382,"y":1830,"type":"Monster"},{"x":1382,"y":1830,"type":"Monster"}],"elementArr":[{"x":480,"y":528,"type":"Door"},{"x":1104,"y":336,"type":"PowerUp"},{"x":912,"y":1776,"type":"Fire"},{"x":960,"y":1728,"type":"Fire"},{"x":1008,"y":1872,"type":"Fire"},{"x":1056,"y":1920,"type":"Fire"}],"npcArr":[{"x":1048,"y":882,"type":"Villager"}]},"playerData":{"spriteCounter":10,"worldX":616,"lives":6,"worldY":2012,"spriteNum":2,"speed":5,"direction":0,"rubies":5}}
