+++ src/main/java/org/project/Objects/ElementHandler.java
@@ -1,7 +1,8 @@
-package org.project;
+package org.project.Objects;
 
 import org.project.Entities.Monster;
 import org.project.Entities.Villager;
+import org.project.GamePanel;
 
 import static org.project.SystemVariables.*;
