+++ src/main/java/org/project/Objects/Fire.java
@@ -1,4 +1,7 @@
-package org.project;
+package org.project.Objects;
+
+import org.project.GamePanel;
+import org.project.Objects.Element;
 
 import javax.imageio.ImageIO;
 import java.awt.*;
