+++ src/test/java/org/project/ElementHandlerTest.java
@@ -5,6 +5,8 @@ import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
 import org.project.Entities.Monster;
 import org.project.Entities.Villager;
+import org.project.Objects.ElementHandler;
+import org.project.Objects.Ruby;
 
 import static org.project.SystemVariables.*;
