+++ src/test/java/org/project/FireTest.java
@@ -1,6 +1,7 @@
 package org.project;
 
 import org.junit.jupiter.api.Test;
+import org.project.Objects.Fire;
 
 import javax.imageio.ImageIO;
 import java.awt.*;
